requirements.txt
django==1.5.1
psycopg2==2.4.1
south==0.7.6

django-tastypie==0.10.0
python-dateutil>=2.1
lxml==3.2.3
defusedxml==0.4.1
PyYAML==3.10
python-mimeparse>=0.1.4
